file based storages

+++ accounts-db/src/append_vec.rs
@@ -15,6 +15,8 @@ use {
         },
         accounts_hash::AccountHash,
         accounts_index::ZeroLamport,
+        buffered_reader::{BufferedReader, BufferedReaderStatus},
+        file_io::read_buffer,
         storable_accounts::StorableAccounts,
         u64_align,
     },
@@ -28,7 +30,7 @@ use {
     },
     std::{
         convert::TryFrom,
-        fs::{remove_file, OpenOptions},
+        fs::{remove_file, File, OpenOptions},
         io::{Seek, SeekFrom, Write},
         mem,
         path::{Path, PathBuf},
@@ -85,7 +87,16 @@ pub enum AppendVecError {
 /// A slice whose contents are known to be valid.
 /// The slice contains no undefined bytes.
 #[derive(Debug, Copy, Clone)]
-struct ValidSlice<'a>(&'a [u8]);
+pub(crate) struct ValidSlice<'a>(&'a [u8]);
+
+impl<'a> ValidSlice<'a> {
+    pub(crate) fn new(data: &'a [u8]) -> Self {
+        Self(data)
+    }
+    fn len(&self) -> usize {
+        self.0.len()
+    }
+}
 
 /// References to account data stored elsewhere. Getting an `Account` requires cloning
 /// (see `StoredAccountMeta::clone_account()`).
@@ -213,6 +224,9 @@ struct AccountOffsets {
 enum AppendVecFileBacking {
     /// A file-backed block of memory that is used to store the data for each appended item.
     MmapOnly(MmapOnly),
+    /// This was opened as a read only file
+    #[allow(dead_code)]
+    FileOnly(File),
 }
 
 #[cfg_attr(feature = "frozen-abi", derive(AbiExample))]
@@ -248,6 +262,9 @@ pub struct AppendVec {
 
     /// The number of bytes available for storing items.
     file_size: u64,
+
+    /// if true, remove file when dropped
+    remove_file_on_drop: AtomicBool,
 }
 
 lazy_static! {
@@ -264,13 +281,16 @@ impl Drop for AppendVec {
                     APPEND_VEC_MMAPPED_FILES_DIRTY.fetch_sub(1, Ordering::Relaxed);
                 }
             }
+            AppendVecFileBacking::FileOnly(_) => {}
         }
-        if let Err(_err) = remove_file(&self.path) {
-            // promote this to panic soon.
-            // disabled due to many false positive warnings while running tests.
-            // blocked by rpc's upgrade to jsonrpc v17
-            //error!("AppendVec failed to remove {}: {err}", &self.path.display());
-            inc_new_counter_info!("append_vec_drop_fail", 1);
+        if self.remove_file_on_drop.load(Ordering::Acquire) {
+            if let Err(_err) = remove_file(&self.path) {
+                // promote this to panic soon.
+                // disabled due to many false positive warnings while running tests.
+                // blocked by rpc's upgrade to jsonrpc v17
+                //error!("AppendVec failed to remove {}: {err}", &self.path.display());
+                inc_new_counter_info!("append_vec_drop_fail", 1);
+            }
         }
     }
 }
@@ -332,6 +352,7 @@ impl AppendVec {
             append_lock: Mutex::new(()),
             current_len: AtomicUsize::new(initial_len),
             file_size: size as u64,
+            remove_file_on_drop: AtomicBool::new(true),
         }
     }
 
@@ -367,6 +388,7 @@ impl AppendVec {
                 }
                 Ok(())
             }
+            AppendVecFileBacking::FileOnly(_file) => Ok(()),
         }
     }
 
@@ -379,9 +401,26 @@ impl AppendVec {
 
     /// when we can use file i/o as opposed to mmap, this is the trigger to tell us
     /// that no more appending will occur and we can close the initial mmap.
+    #[allow(dead_code)]
     pub(crate) fn reopen_as_readonly(&self) -> Option<Self> {
-        // this is a no-op when we are already a mmap
-        None
+        #[cfg(not(unix))]
+        // must open as mmmap on non-unix
+        return None;
+
+        #[cfg(unix)]
+        match &self.backing {
+            // already a file, so already read-only
+            AppendVecFileBacking::FileOnly(_file) => None,
+            AppendVecFileBacking::MmapOnly(_mmap) => {
+                // we are a map, so re-open as a file
+                _ = self.flush();
+                // we are re-opening the file, so don't remove the file on disk when the old mmapped one is dropped
+                self.remove_file_on_drop.store(false, Ordering::Release);
+                AppendVec::new_from_file(self.path.clone(), self.len(), StorageAccess::File)
+                    .ok()
+                    .map(|(av, _size)| av)
+            }
+        }
     }
 
     /// how many more bytes can be stored in this append vec
@@ -423,10 +462,11 @@ impl AppendVec {
     }
 
     /// Creates an appendvec from file without performing sanitize checks or counting the number of accounts
+    #[allow(unused_variables)]
     pub fn new_from_file_unchecked(
         path: impl Into<PathBuf>,
         current_len: usize,
-        _storage_access: StorageAccess,
+        storage_access: StorageAccess,
     ) -> Result<Self> {
         let path = path.into();
         let file_size = std::fs::metadata(&path)?.len();
@@ -437,6 +477,21 @@ impl AppendVec {
             .write(true)
             .create(false)
             .open(&path)?;
+
+        #[cfg(unix)]
+        // we must use mmap on non-linux
+        if storage_access == StorageAccess::File {
+            APPEND_VEC_MMAPPED_FILES_OPEN.fetch_add(1, Ordering::Relaxed);
+
+            return Ok(AppendVec {
+                path,
+                backing: AppendVecFileBacking::FileOnly(data),
+                append_lock: Mutex::new(()),
+                current_len: AtomicUsize::new(current_len),
+                file_size,
+                remove_file_on_drop: AtomicBool::new(true),
+            });
+        }
 
         let mmap = unsafe {
             let result = MmapMut::map_mut(&data);
@@ -457,6 +512,7 @@ impl AppendVec {
             append_lock: Mutex::new(()),
             current_len: AtomicUsize::new(current_len),
             file_size,
+            remove_file_on_drop: AtomicBool::new(true),
         })
     }
 
@@ -529,6 +585,9 @@ impl AppendVec {
                 };
                 *offset = pos + len;
             }
+            AppendVecFileBacking::FileOnly(_file) => {
+                unimplemented!();
+            }
         }
     }
 
@@ -598,6 +657,60 @@ impl AppendVec {
                         hash,
                     },
                 )))
+            }
+            AppendVecFileBacking::FileOnly(file) => {
+                // 4096 was just picked to be a single page size
+                let mut buf = [0u8; 4096];
+                if let Ok(bytes_read) = read_buffer(file, offset, &mut buf, self.len()) {
+                    let valid_bytes = ValidSlice(&buf[..bytes_read]);
+                    let (meta, next): (&StoredMeta, _) = Self::get_type(valid_bytes, 0)?;
+                    let (account_meta, next): (&AccountMeta, _) =
+                        Self::get_type(valid_bytes, next)?;
+                    let (hash, next): (&AccountHash, _) = Self::get_type(valid_bytes, next)?;
+                    let data_len = meta.data_len;
+                    let remaining_bytes_for_data = bytes_read - next;
+                    if remaining_bytes_for_data >= data_len as usize {
+                        // we already read enough data to load this account
+                        let (data, next) =
+                            Self::get_slice(valid_bytes, next, meta.data_len as usize)?;
+                        let stored_size = next;
+                        let account = StoredAccountMeta::AppendVec(AppendVecStoredAccountMeta {
+                            meta,
+                            account_meta,
+                            data,
+                            offset,
+                            stored_size,
+                            hash,
+                        });
+                        return Some(callback(account));
+                    } else {
+                        // not enough was read from file to get `data`
+                        assert!(data_len <= 20_000_000, "{}", data_len);
+                        let mut data = vec![0u8; data_len as usize];
+                        // instead, we could piece together what we already read here. Maybe we just needed 1 more byte.
+                        // Note here `next` is a 0-based offset from the beginning of this account.
+                        let Ok(bytes_read) =
+                            read_buffer(file, offset + next, &mut data, self.len())
+                        else {
+                            return None;
+                        };
+                        if bytes_read < data_len as usize {
+                            // eof or otherwise couldn't read all the data
+                            return None;
+                        }
+                        let stored_size = aligned_stored_size(data_len as usize);
+                        let account = StoredAccountMeta::AppendVec(AppendVecStoredAccountMeta {
+                            meta,
+                            account_meta,
+                            data: &data[..],
+                            offset,
+                            stored_size,
+                            hash,
+                        });
+                        return Some(callback(account));
+                    }
+                }
+                None
             }
         }
     }
@@ -718,26 +831,91 @@ impl AppendVec {
                     offset = next.next_account_offset;
                 }
             }
+            AppendVecFileBacking::FileOnly(file) => {
+                let mut reader =
+                    BufferedReader::new(1024 * 1024 * 32, self.len(), file, STORE_META_OVERHEAD);
+                while reader.read().ok() == Some(BufferedReaderStatus::Success) {
+                    let (offset, bytes_subset) = reader.get_data_and_offset();
+                    let (meta, next): (&StoredMeta, _) = Self::get_type(bytes_subset, 0).unwrap();
+                    let (account_meta, next): (&AccountMeta, _) =
+                        Self::get_type(bytes_subset, next).unwrap();
+                    let (_hash, next): (&AccountHash, _) =
+                        Self::get_type(bytes_subset, next).unwrap();
+                    let stored_size_aligned = u64_align!(next + (meta.data_len as usize));
+                    callback(IndexInfo {
+                        index_info: {
+                            IndexInfoInner {
+                                pubkey: meta.pubkey,
+                                lamports: account_meta.lamports,
+                                offset,
+                                data_len: meta.data_len,
+                                executable: account_meta.executable,
+                                rent_epoch: account_meta.rent_epoch,
+                            }
+                        },
+                        stored_size_aligned,
+                    });
+                    reader.advance_offset(stored_size_aligned);
+                }
+            }
         }
     }
 
     /// Iterate over all accounts and call `callback` with each account.
     #[allow(clippy::blocks_in_conditions)]
     pub fn scan_accounts(&self, mut callback: impl for<'local> FnMut(StoredAccountMeta<'local>)) {
-        let mut offset = 0;
-        while self
-            .get_stored_account_meta_callback(offset, |account| {
-                offset += account.stored_size();
-                if account.is_zero_lamport() && account.pubkey() == &Pubkey::default() {
-                    // we passed the last useful account
-                    return false;
-                }
+        match &self.backing {
+            AppendVecFileBacking::MmapOnly(_mmap) => {
+                let mut offset = 0;
+                while self
+                    .get_stored_account_meta_callback(offset, |account| {
+                        offset += account.stored_size();
+                        if account.is_zero_lamport() && account.pubkey() == &Pubkey::default() {
+                            // we passed the last useful account
+                            return false;
+                        }
 
-                callback(account);
-                true
-            })
-            .unwrap_or_default()
-        {}
+                        callback(account);
+                        true
+                    })
+                    .unwrap_or_default()
+                {}
+            }
+            AppendVecFileBacking::FileOnly(file) => {
+                // big enough for > 2x the largest account
+                let mut reader =
+                    BufferedReader::new(1024 * 1024 * 32, self.len(), file, STORE_META_OVERHEAD);
+                while reader.read().ok() == Some(BufferedReaderStatus::Success) {
+                    let (offset, bytes_subset) = reader.get_data_and_offset();
+                    let (meta, next): (&StoredMeta, _) = Self::get_type(bytes_subset, 0).unwrap();
+                    let (account_meta, next): (&AccountMeta, _) =
+                        Self::get_type(bytes_subset, next).unwrap();
+                    let (hash, next): (&AccountHash, _) =
+                        Self::get_type(bytes_subset, next).unwrap();
+                    let data_len = meta.data_len;
+                    if bytes_subset.len() - next >= data_len as usize {
+                        // we already read enough data to load this account
+                        let data = &bytes_subset.0[next..(next + data_len as usize)];
+                        let stored_size = u64_align!(next + (data_len as usize));
+                        let account = StoredAccountMeta::AppendVec(AppendVecStoredAccountMeta {
+                            meta,
+                            account_meta,
+                            data,
+                            offset,
+                            stored_size,
+                            hash,
+                        });
+                        callback(account);
+                        reader.advance_offset(stored_size);
+                    } else {
+                        // fall through and read the whole account again. we need refs for StoredMeta and data.
+                        reader.set_required_data_len(
+                            STORE_META_OVERHEAD.saturating_add(data_len as usize),
+                        )
+                    }
+                }
+            }
+        }
     }
 
     /// for each offset in `sorted_offsets`, get the size of the account. No other information is needed for the account.
@@ -756,6 +934,13 @@ impl AppendVec {
                         break;
                     }
                     result.push(next.stored_size_aligned);
+                }
+            }
+            AppendVecFileBacking::FileOnly(_file) => {
+                for &offset in sorted_offsets {
+                    self.get_stored_account_meta_callback(offset, |stored_meta| {
+                        result.push(stored_meta.stored_size());
+                    });
                 }
             }
         }
@@ -785,6 +970,11 @@ impl AppendVec {
                     callback(&stored_meta.pubkey);
                     offset = next.next_account_offset;
                 }
+            }
+            AppendVecFileBacking::FileOnly(_file) => {
+                self.scan_accounts(|stored_meta| {
+                    callback(stored_meta.pubkey());
+                });
             }
         }
     }
@@ -861,6 +1051,7 @@ impl AppendVec {
                     }
                 }
             }
+            AppendVecFileBacking::FileOnly(_) => {}
         }
 
         (!offsets.is_empty()).then(|| {
@@ -876,15 +1067,17 @@ impl AppendVec {
         })
     }
 
-    /// true if this storage can possibly be appended to (independent of capacity check)
     pub(crate) fn can_append(&self) -> bool {
-        // always can append to a mmapped append vec
-        true
+        match &self.backing {
+            AppendVecFileBacking::FileOnly(_file) => false,
+            AppendVecFileBacking::MmapOnly(_mmap) => true,
+        }
     }
 
     /// Returns the way to access this accounts file when archiving
     pub(crate) fn internals_for_archive(&self) -> InternalsForArchive {
         match &self.backing {
+            AppendVecFileBacking::FileOnly(_file) => InternalsForArchive::FileIo(self.path()),
             // note this returns the entire mmap slice, even bytes that we consider invalid
             AppendVecFileBacking::MmapOnly(MmapOnly { mmap, .. }) => {
                 InternalsForArchive::Mmap(mmap)
@@ -1014,6 +1207,7 @@ pub mod tests {
     }
 
     #[test_case(StorageAccess::Mmap)]
+    #[test_case(StorageAccess::File)]
     fn test_append_vec_new_from_file_bad_size(storage_access: StorageAccess) {
         let file = get_append_vec_path("test_append_vec_new_from_file_bad_size");
         let path = &file.path;
@@ -1259,6 +1453,7 @@ pub mod tests {
     }
 
     #[test_case(StorageAccess::Mmap)]
+    #[test_case(StorageAccess::File)]
     fn test_new_from_file_crafted_zero_lamport_account(storage_access: StorageAccess) {
         // This test verifies that when we sanitize on load, that we fail sanitizing if we load an account with zero lamports that does not have all default value fields.
         // This test writes an account with zero lamports, but with 3 bytes of data. On load, it asserts that load fails.
@@ -1331,6 +1526,7 @@ pub mod tests {
     }
 
     #[test_case(StorageAccess::Mmap)]
+    #[test_case(StorageAccess::File)]
     fn test_new_from_file_crafted_data_len(storage_access: StorageAccess) {
         let file = get_append_vec_path("test_new_from_file_crafted_data_len");
         let path = &file.path;
@@ -1363,6 +1559,7 @@ pub mod tests {
     }
 
     #[test_case(StorageAccess::Mmap)]
+    #[test_case(StorageAccess::File)]
     fn test_new_from_file_too_large_data_len(storage_access: StorageAccess) {
         let file = get_append_vec_path("test_new_from_file_too_large_data_len");
         let path = &file.path;
@@ -1396,6 +1593,7 @@ pub mod tests {
     }
 
     #[test_case(StorageAccess::Mmap)]
+    #[test_case(StorageAccess::File)]
     fn test_new_from_file_crafted_executable(storage_access: StorageAccess) {
         let file = get_append_vec_path("test_new_from_crafted_executable");
         let path = &file.path;
